# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type WorkoutType = typeof WorkoutType[keyof typeof WorkoutType];


export const WorkoutType = {
  strength: 'strength',
  cardio: 'cardio',
  flexibility: 'flexibility',
  hiit: 'hiit',
  sport: 'sport',
  other: 'other',
} as const;

export interface Workout {
  id: number;
  name: string;
  date: string;
  type: WorkoutType;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
  exerciseCount?: number;
  createdAt: string;
}

export type WorkoutDetailType = typeof WorkoutDetailType[keyof typeof WorkoutDetailType];


export const WorkoutDetailType = {
  strength: 'strength',
  cardio: 'cardio',
  flexibility: 'flexibility',
  hiit: 'hiit',
  sport: 'sport',
  other: 'other',
} as const;

export interface Exercise {
  id: number;
  workoutId: number;
  name: string;
  muscleGroup: string;
  /** @nullable */
  sets?: number | null;
  /** @nullable */
  reps?: number | null;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
  order: number;
}

export interface WorkoutDetail {
  id: number;
  name: string;
  date: string;
  type: WorkoutDetailType;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
  exercises: Exercise[];
  createdAt: string;
}

export type WorkoutInputType = typeof WorkoutInputType[keyof typeof WorkoutInputType];


export const WorkoutInputType = {
  strength: 'strength',
  cardio: 'cardio',
  flexibility: 'flexibility',
  hiit: 'hiit',
  sport: 'sport',
  other: 'other',
} as const;

export interface WorkoutInput {
  /** @minLength 1 */
  name: string;
  date: string;
  type: WorkoutInputType;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
}

export type WorkoutUpdateType = typeof WorkoutUpdateType[keyof typeof WorkoutUpdateType];


export const WorkoutUpdateType = {
  strength: 'strength',
  cardio: 'cardio',
  flexibility: 'flexibility',
  hiit: 'hiit',
  sport: 'sport',
  other: 'other',
} as const;

export interface WorkoutUpdate {
  /** @minLength 1 */
  name?: string;
  date?: string;
  type?: WorkoutUpdateType;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface ExerciseInput {
  /** @minLength 1 */
  name: string;
  muscleGroup: string;
  /** @nullable */
  sets?: number | null;
  /** @nullable */
  reps?: number | null;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface ExerciseUpdate {
  name?: string;
  muscleGroup?: string;
  /** @nullable */
  sets?: number | null;
  /** @nullable */
  reps?: number | null;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
}

export type ExerciseTemplateType = typeof ExerciseTemplateType[keyof typeof ExerciseTemplateType];


export const ExerciseTemplateType = {
  strength: 'strength',
  cardio: 'cardio',
  flexibility: 'flexibility',
  hiit: 'hiit',
  sport: 'sport',
  other: 'other',
} as const;

export interface ExerciseTemplate {
  id: number;
  name: string;
  muscleGroup: string;
  type: ExerciseTemplateType;
}

export interface StatsSummary {
  totalWorkouts: number;
  totalDurationMinutes: number;
  thisWeekWorkouts: number;
  thisMonthWorkouts: number;
  avgDurationMinutes: number;
  currentStreak: number;
}

export interface WeeklyStats {
  week: string;
  count: number;
  totalDurationMinutes: number;
}

export interface MuscleGroupStats {
  muscleGroup: string;
  count: number;
}

export type ListWorkoutsParams = {
limit?: number;
offset?: number;
};

export type GetWeeklyStatsParams = {
weeks?: number;
};