# lib/api-zod/src/generated/types/exercise.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Exercise {
  id: number;
  workoutId: number;
  name: string;
  muscleGroup: string;
  /** @nullable */
  sets?: number | null;
  /** @nullable */
  reps?: number | null;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
  order: number;
}