# lib/api-zod/src/generated/types/exerciseTemplate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { ExerciseTemplateType } from './exerciseTemplateType';

export interface ExerciseTemplate {
  id: number;
  name: string;
  muscleGroup: string;
  type: ExerciseTemplateType;
}