# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalWorkouts: number;
  totalDurationMinutes: number;
  thisWeekWorkouts: number;
  thisMonthWorkouts: number;
  avgDurationMinutes: number;
  currentStreak: number;
}