# lib/api-zod/src/generated/types/weeklyStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface WeeklyStats {
  week: string;
  count: number;
  totalDurationMinutes: number;
}