# lib/api-zod/src/generated/types/workout.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkoutType } from './workoutType';

export interface Workout {
  id: number;
  name: string;
  date: Date;
  type: WorkoutType;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
  exerciseCount?: number;
  createdAt: Date;
}