# lib/api-zod/src/generated/types/workoutDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { Exercise } from './exercise';
import type { WorkoutDetailType } from './workoutDetailType';

export interface WorkoutDetail {
  id: number;
  name: string;
  date: Date;
  type: WorkoutDetailType;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
  exercises: Exercise[];
  createdAt: Date;
}