# lib/api-zod/src/generated/types/workoutInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkoutInputType } from './workoutInputType';

export interface WorkoutInput {
  /** @minLength 1 */
  name: string;
  date: Date;
  type: WorkoutInputType;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
}