# lib/api-zod/src/generated/types/workoutUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fitness Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkoutUpdateType } from './workoutUpdateType';

export interface WorkoutUpdate {
  /** @minLength 1 */
  name?: string;
  date?: Date;
  type?: WorkoutUpdateType;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
}